fix: to(torch.int64)

+++ lit_model.py
@@ -66,7 +66,8 @@ class LitModel(L.LightningModule):
         )
 
         self.register_buffer(
-            "current_hard_prompt", tokenizer.encode("✅579 * 640 = 370,560✅", bos=False)
+            "current_hard_prompt",
+            tokenizer.encode("✅579 * 640 = 370,560✅", bos=False).to(torch.int64),
         )
 
     def training_step(self, batch: dict, batch_idx: int) -> torch.Tensor:
